removed user and book bundles, added user is logged in filter post function
