+++ app/Resources/views/base.html.twig
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta charset="UTF-8" />
-        <title>{% block title %}symblog{% endblock %} - symblog</title>
+        <title>{% block title %}Rate a book{% endblock %}</title>
         {% block stylesheets %}
             {% stylesheets
                 'https://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css'
@@ -36,7 +36,7 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a class="navbar-brand" href="#">Symblog</a>
+                <a class="navbar-brand" href="#">Rate a book</a>
                 <div class="collapse navbar-collapse">
                     <ul class="nav navbar-nav navbar-right">
                         <li><a href="{{ path('index') }}">Home</a></li>
@@ -64,7 +64,7 @@
         <br>
         <div class="footer">
             {% block footer %}
-                <p>Symfony blog tutorial</p>
+                <p>Rate a book</p>
             {% endblock %} </div>
     </div>
     </body>
