+++ src/Blogger/BlogBundle/Repository/BookRepository.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Blogger\UserBundle\Repository;
+namespace Blogger\BlogBundle\Repository;
 
 /**
- * UserRepository
+ * BookRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class UserRepository extends \Doctrine\ORM\EntityRepository
+class BookRepository extends \Doctrine\ORM\EntityRepository
 {
 }
