+++ src/Blogger/BlogBundle/Repository/UserRepository.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Blogger\BookBundle\Repository;
+namespace Blogger\BlogBundle\Repository;
 
 /**
- * BookRepository
+ * UserRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class BookRepository extends \Doctrine\ORM\EntityRepository
+class UserRepository extends \Doctrine\ORM\EntityRepository
 {
 }
