--- src/Blogger/BookBundle/BookBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Blogger\BookBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class BookBundle extends Bundle
-{
-}
