--- src/Blogger/BookBundle/Resources/config/services.yml
@@ -1,4 +1,0 @@
-services:
-#    book.example:
-#        class: BookBundle\Example
-#        arguments: ["@service_id", "plain_value", "%parameter%"]
