--- src/Blogger/UserBundle/Controller/DefaultController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace UserBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-class DefaultController extends Controller
-{
-    public function indexAction()
-    {
-        return $this->render('UserBundle:Default:index.html.twig');
-    }
-}
