--- src/Blogger/UserBundle/Resources/config/routing.yml
@@ -1,7 +1,0 @@
-user_login:
-    path:     /login
-    defaults: { _controller: FOSUserBundle:Security:login }
-
-user_logout:
-    path:     /logout
-    defaults: { _controller: FOSUserBundle:Security:logout }
