--- src/Blogger/UserBundle/Resources/config/services.yml
@@ -1,4 +1,0 @@
-services:
-#    user.example:
-#        class: UserBundle\Example
-#        arguments: ["@service_id", "plain_value", "%parameter%"]
