--- src/Blogger/UserBundle/UserBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace UserBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class UserBundle extends Bundle
-{
-}
